refactor: move `AiProviders` to `@activepieces/pieces-common`
