+++ packages/pieces/community/common/src/lib/ai/providers/anthropic/index.ts
@@ -16,7 +16,6 @@ export const anthropic: AIFactory = ({
     },
   })
   return {
-    underlying: sdk,
     provider: "ANTHROPIC" as const,
     chat: {
       text: async (params) => {
