+++ packages/pieces/community/common/src/lib/ai/providers/openai/index.ts
@@ -15,7 +15,6 @@ export const openai: AIFactory = ({
     },
   });
   return {
-    underlying: sdk,
     provider: "OPENAI",
     chat: {
       text: async (params) => {
