+++ packages/react-ui/src/app/routes/platform/projects/index.tsx
@@ -148,7 +148,7 @@ export default function ProjectsPage() {
               cell: ({ row }) => {
                 return (
                   <div className="text-left">
-                    {formatUtils.formatNumber(row.original.usage.aiTokensUsage)} /{' '}
+                    {formatUtils.formatNumber(row.original.usage.aiTokens)} /{' '}
                     {row.original.plan.aiTokens ? formatUtils.formatNumber(row.original.plan.aiTokens) : '-'}
                   </div>
                 );
