+++ packages/react-ui/tsconfig.app.json
@@ -23,6 +23,9 @@
       "@activepieces/pieces-framework": [
         "../../packages/pieces/community/framework/src" 
       ],
+      "@activepieces/pieces-common": [
+        "../../packages/pieces/community/common/src" 
+      ],
       "ee-embed-sdk": [
         "../../packages/ee/ui/embed-sdk/src"
       ]
